voice/Live samples: libxvoice wiring + missing XDK files + MSVC-ism fixes (32/37 build)
- RxdkLibraries += libxvoice on the 23 samples whose XbVoice/XHV code
  references the voice device tables or engine; += libxact on the UIX
  samples with XACT audio; += libxonline/libxnet where the secure-socket
  path (WSA*/XNet*/CXoBase kerb) is linked.
- Files the 5849 XDK install itself omits, restored from siblings:
  UIXPlayers' XHVVoiceManager.cpp/h (from UIXFriends), Signin.cpp for
  SimpleMessaging/SimpleConnect (from SimpleFriends) and SimpleTeams
  (adapted to its 3-arg service-list SignIn), and a hand-generated
  SampleSRBank_en.h for SpeechRecognition (the srbld tool is not part of
  RXDK; vocab IDs follow the .txt section order).
- Arbitration Common.h: #include "match.h" -> MatchMaking.h (no match.h
  exists anywhere in the XDK; the sample ships MatchMaking.h).
- MSVC-isms clang rejects: implicit-int params (Integrated/Storage
  UserContent 'const dwSigType', Integrated SCORE_STRING_SIZE), lvalue
  casts (Marketplace 'packet >> (WORD)w'), old for-scope leaks
  (Stats.cpp, HostMigration.cpp), friend decls not injecting names +
  __inline members defined in another TU (LowLevelVoiceChat), and a
  missing <wctype.h> for towupper (Common XbRandName.cpp).

With RXDK-Libs 244bf0c: all 9 voice samples and 32/37 of the
Networking+XboxLive3 set build to bootable .xbe. Remaining 5: Arbitration
(ATG OptiMatch framework types), ContentDownload+SimpleContentDownload
(char_traits<unsigned char>), UIXKeyboard+UIXPlugin (sk_res.h skin build).

+++ XDKSamples/Networking/Marketplace/marketplace.cpp
@@ -647,12 +647,12 @@ HRESULT Marketplace::FrameMove()
             // If message waiting, process it
             if( iResult != SOCKET_ERROR && iResult > 0 )
             {
-                packet >> (WORD)wMessage;
+                packet >> wMessage;
                 switch (wMessage)
                 {
                 case MSG_JOIN_OK:               
                     m_State = STATE_PLAYING_GAME;
-                    packet >> (WORD)wMe;                   
+                    packet >> wMe;                   
                     
                     g_PlayerMgr.ReadPlayerListFull( packet );
                     g_PlayerMgr.SetLocalPlayerLUID( wMe );
